rest-api/requirements.txt
Flask
earthengine-api
google-cloud-firestore
google-cloud-pubsub
google-cloud-storage
mercantile
oauth2client
requests